Migrate to the new cy config
