--- cypress.json
@@ -1,4 +1,0 @@
-{
-  "fixturesFolder": false,
-  "supportFile": false
-}
